Removed The "Static Get" Methods
Removed the "Static Get" methods because originally the get methods were dynamic dispatch and polymorphic VI's didn't support dynamic dispatch so I had to add a Static call in between. The reason for the "Get"'s to be dynamic dispatch has been removed and now they are static dispatch, this means the "Static Gets" are redundant.